+++ scripts/projs/app.lua
@@ -6,4 +6,8 @@ end
 
 function App:setup_project()
 	kind('ConsoleApp')
+	postbuildcommands{
+		self.registered_dll,
+		'ver >nul', -- Ignore robocopy errors
+	}
 end